requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai
chromadb
pydantic
